Modularize the code used to display embedded media
